requirements.txt
cffi==1.14.1
click==7.1.2
cryptography==3.0
Flask==1.1.2
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.11.2
mariadb==1.0.0
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
